refactor(jsonrpc): move ID generation to dedicated file and update generator
