+++ jsonrpc/types.go
@@ -39,78 +39,3 @@ func (e *Error) Error() string {
 func (e *Error) Unwrap() error {
 	return errors.New(e.Message)
 }
-
-type ID struct {
-	str   *string
-	num   *float64
-	isNil bool
-}
-
-func (i *ID) String() string {
-	if i.isNil {
-		return "null"
-	}
-	if i.str != nil {
-		return *i.str
-	}
-	if i.num != nil {
-		return fmt.Sprintf("%v", *i.num)
-	}
-	return "null"
-}
-
-type id interface {
-	~string | ~float64 | ~int
-}
-
-func NewID[T id](v T) *ID {
-	var id ID
-	switch any(v).(type) {
-	case string:
-		s := any(v).(string)
-		id.str = &s
-	case float64:
-		n := any(v).(float64)
-		id.num = &n
-	case int:
-		n := float64(any(v).(int))
-		id.num = &n
-	default:
-		return nil
-	}
-	return &id
-}
-
-func (i *ID) UnmarshalJSON(data []byte) error {
-	if string(data) == "null" {
-		i.isNil = true
-		return nil
-	}
-
-	var s string
-	if err := json.Unmarshal(data, &s); err == nil {
-		i.str = &s
-		return nil
-	}
-
-	var n float64
-	if err := json.Unmarshal(data, &n); err == nil {
-		i.num = &n
-		return nil
-	}
-
-	return nil
-}
-
-func (i ID) MarshalJSON() ([]byte, error) {
-	if i.isNil {
-		return []byte("null"), nil
-	}
-	if i.str != nil {
-		return json.Marshal(i.str)
-	}
-	if i.num != nil {
-		return json.Marshal(i.num)
-	}
-	return []byte("null"), nil
-}
